// backend/ledger/src/features/gitea/client/gitea-api.ts
export interface Commit {
  /** User represents a user */
  author?: User;
  commit?: RepoCommit;
  /** User represents a user */
  committer?: User;
  /** @format date-time */
  created?: string;
  files?: CommitAffectedFiles[];
  html_url?: string;
  parents?: CommitMeta[];
  sha?: string;
  /** CommitStats is statistics for a RepoCommit */
  stats?: CommitStats;
  url?: string;
}

/** CommitAffectedFiles store information about files affected by the commit */
export interface CommitAffectedFiles {
  filename?: string;
  status?: string;
}

/** CommitDateOptions store dates for GIT_AUTHOR_DATE and GIT_COMMITTER_DATE */
export interface CommitDateOptions {
  /** @format date-time */
  author?: string;
  /** @format date-time */
  committer?: string;
}

/** CommitMeta contains meta information of a commit in terms of API. */
export interface CommitMeta {
  /** @format date-time */
  created?: string;
  sha?: string;
  url?: string;
}

/** CommitStats is statistics for a RepoCommit */
export interface CommitStats {
  /** @format int64 */
  additions?: number;
  /** @format int64 */
  deletions?: number;
  /** @format int64 */
  total?: number;
}
// ...
export interface CommitUser {
  date?: string;
  /** @format email */
  email?: string;
  name?: string;
}
// ...
export interface PayloadCommitVerification {
  payload?: string;
  reason?: string;
  signature?: string;
  /** PayloadUser represents the author or committer of a commit */
  signer?: PayloadUser;
  verified?: boolean;
}
// ...
export interface RepoCommit {
  author?: CommitUser;
  committer?: CommitUser;
  message?: string;
  tree?: CommitMeta;
  url?: string;
  /** PayloadCommitVerification represents the GPG verification of a commit */
  verification?: PayloadCommitVerification;
}
// ...
export interface User {
  /** Is user active */
  active?: boolean;
  /** URL to the user's avatar */
  avatar_url?: string;
  /** @format date-time */
  created?: string;
  /** the user's description */
  description?: string;
  /** @format email */
  email?: string;
  /**
   * user counts
   * @format int64
   */
  followers_count?: number;
  /** @format int64 */
  following_count?: number;
  /** the user's full name */
  full_name?: string;
  /** URL to the user's gitea page */
  html_url?: string;
  /**
   * the user's id
   * @format int64
   */
  id?: number;
  /** Is the user an administrator */
  is_admin?: boolean;
  /** User locale */
  language?: string;
  /** @format date-time */
  last_login?: string;
  /** the user's location */
  location?: string;
  /** the user's username */
  login?: string;
  /**
   * the user's authentication sign-in name.
   * @default "empty"
   */
  login_name?: string;
  /** Is user login prohibited */
  prohibit_login?: boolean;
  /** Is user restricted */
  restricted?: boolean;
  /**
   * The ID of the user's Authentication Source
   * @format int64
   */
  source_id?: number;
  /** @format int64 */
  starred_repos_count?: number;
  /** User visibility level option: public, limited, private */
  visibility?: string;
  /** the user's website */
  website?: string;
}